requirements.txt
aiosqlite==0.17.0
asgiref==3.4.1
asyncpg==0.24.0
click==8.0.1
fastapi==0.68.1
h11==0.12.0
iso8601==0.1.16
passlib==1.7.4
Pillow==7.2.0
pydantic==1.8.2
PyJWT==2.1.0
pypika-tortoise==0.1.1
python-decouple==3.4
pytz==2021.1
starlette==0.14.2
tortoise-orm==0.17.7
typing-extensions==3.10.0.2
uvicorn==0.15.0
